chore: modularize map renderers and utils
